OC/P11/requirements.txt
Pandas
pillow
tensorflow
pyspark
pyarrow
ipykernel
matplotlib